social_media_api/requirements.txt
django>=5.0
djangorestframework
django-filter
django-environ
gunicorn
whitenoise
dj-database-url
psycopg2-binary
pillow
